bugs fixed

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -2,9 +2,9 @@ package lang.print.gaps.task5;
 
 public class NumberDivider {
     void divide(int toBeDivided) {
-        double divided = toBeDivided / 5.0;
+        double divided = (double) toBeDivided / 5;
     }
 
 
-
 }
+
